requirements.txt
Django==1.6.5
Pillow==2.5.1
PyYAML==3.11
South==1.0
-e git+https://github.com/djacobs/PyAPNs.git@e2666a52b6bde42bb149ef944017edf49927a032#egg=apns-dev
defusedxml==0.4.1
dj-database-url==0.3.0
dj-static==0.0.5
django-oauth2-provider==0.2.6.1
django-tastypie==0.11.1
django-toolbelt==0.0.1
gunicorn==19.0.0
lxml==3.3.5
psycopg2==2.5.3
pystache==0.5.4
python-dateutil==2.2
python-mimeparse==0.1.4
shortuuid==0.4.2
six==1.7.3
static==1.0.2
wsgiref==0.1.2
